requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.18.29
botocore==1.21.29
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
decorator==5.0.9
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.5
gunicorn==20.1.0
hvac==0.11.0
idna==3.2
image==1.5.33
importlib-metadata==4.6.4
Jinja2==3.0.1
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
paramiko==2.7.2
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
sweetify==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
validators==0.18.2
whitenoise==5.3.0
win32-setctime==1.0.3
zipp==3.5.0
